requirements.txt
CacheControl==0.12.5
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
dialogflow==0.6.0
firebase==3.0.1
firebase-admin==2.18.0
Flask==1.0.2
future==0.17.1
get==2019.4.13
google-api-core==1.14.2
google-api-python-client==1.7.11
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud==0.34.0
google-cloud-core==1.0.3
google-cloud-firestore==1.4.0
google-cloud-storage==1.19.0
google-resumable-media==0.4.0
googleapis-common-protos==1.6.0
grpcio==1.23.0
gunicorn==19.9.0
httplib2==0.13.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
line-bot-sdk==1.14.0
MarkupSafe==1.1.1
msgpack==0.6.1
post==2019.4.13
protobuf==3.9.1
public==2019.4.13
pyasn1==0.4.7
pyasn1-modules==0.2.6
pytz==2019.2
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
rsa==4.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.25.3
Werkzeug==0.15.5